- Payment Options Rework - Logo files added - Admin - Attendee management adjusted

+++ eventplus/app/views/admin/payments/manage.php
@@ -90,29 +90,37 @@ $event_id = $oEvent->id;
                         <td>
                             <?php if (count($payments) > 0): ?>
                                 <?php
+                                $totalPaid = 0;
                                 foreach ($payments as $paymentRow):
 
                                     $deleteLink = $this->adminUrl('admin_payments/delete', array('event_id' => $event_id, 'id' => $paymentRow['id']));
                                     $editLink = $this->adminUrl('admin_payments/edit', array('event_id' => $event_id, 'id' => $paymentRow['id']));
+                                    $viewLink = $this->adminUrl('admin_payments/view', array('event_id' => $event_id, 'id' => $paymentRow['id']));
 
+                                    $totalPaid = $totalPaid + $paymentRow['mc_gross'];
                                     echo $paymentRow['mc_currency'] . " " . $paymentRow['mc_gross'] . " " . $paymentRow['txn_type'] . " " . $paymentRow['txn_id'] . " (" . $paymentRow['payment_date'] . ")" . "     ";
                                     ?>
+                                    <?php if (EventPlus_Models_Payments::isValidMethod($paymentRow['txn_type']) == false): ?>
+                                        <a href="<?php echo $deleteLink; ?>" onclick="return confirm('Are you sure you wish to delete?');"><?php _e('Delete', 'evrplus_language'); ?></a> | 
+                                        <a href="<?php echo $editLink; ?>"><?php _e('Edit', 'evrplus_language'); ?></a> <br />
 
-                                    <a href="<?php echo $deleteLink; ?>" onclick="return confirm('Are you sure you wish to delete?');"><?php _e('Delete', 'evrplus_language'); ?></a> | 
-                                    <a href="<?php echo $editLink; ?>"><?php _e('Edit', 'evrplus_language'); ?></a> <br />
+                                    <?php endif; ?>
+
                                 <?php endforeach; ?>
                             <?php else: ?>
                                 <?php
-                                echo '<font color="red">';
-                                _e('No Payments Received!', 'evrplus_language');
-                                echo '</font>';
+                                echo '<font color="red">' . _e('No Payments Received!', 'evrplus_language') . '</font>';
                                 ?>
                             <?php endif; ?>
                         </td>
                         <td>
 
-                            <a href="<?php echo $this->adminUrl('admin_payments/add', array('event_id' => $event_id, 'attendee_id' => $attendee_id)); ?>" class="button-primary"><?php _e('Add Payment', 'evrplus_language'); ?></a>
-
+                            <?php if ($totalPaid != $payment): ?>
+                                <a href="<?php echo $this->adminUrl('admin_payments/add', array('event_id' => $event_id, 'attendee_id' => $attendee_id)); ?>" class="btn btn-small btn-warning"><?php _e('Add Payment', 'evrplus_language'); ?></a>
+                            <?php else: ?>
+                                 <a href="<?php echo $viewLink; ?>" class="btn btn-small btn-info"><?php _e('View Details', 'evrplus_language'); ?></a> 
+                            <?php endif; ?>
+                                
                         </td>
                     </tr>
